% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pactual function.r
\name{pactual}
\alias{pactual}
\title{Calculate pactual}
\usage{
pactual(ns, pmax = 1, pcrit = 0.05, slices = 1000, tails = 2,
  indent = "")
}
\arguments{
\item{ns}{ns refers to a vector of sample sizes that indicate the initial number of participants run (N1) followed by the number of participants in the first round of augmentation (N2), the number of participants in the second round of augmentation (N3), etc.}

\item{pmax}{refers to the maximum p-value in the data collected thusfar such that the researcher would augment the sample with additional participants.}

\item{pcrit}{refers to the value for determining statistical significance (typically .05).}

\item{slices}{refers to the number of slices to divide the probability distribution (higher numbers of slices will make the calculations slower but more precise).}

\item{tails}{refers to whether the tests are one- or two-tailed.}

\item{indent}{is used to format the countdown when this function is called from pcrit.}
}
\description{
This function calculates pactual, a statistic discussed in Sagarin, Ambler, and Lee's (2014, Perspectives on Psychological Science, 9, 293-304) "An Ethical Approach to Peeking at Data".
}
\examples{
pactual(c(100,50,25), pmax = .1, pcrit = .05, slices = 1000, tails = 1)
pactual()
}
