% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcrit function.r
\name{pcrit}
\alias{pcrit}
\title{Calculate pcrit}
\usage{
pcrit(ns, pmax = 1, pdesired = 0.05, slices = 1000, tails = 2)
}
\arguments{
\item{ns}{refers to a vector of sample sizes that indicate the initial number of participants run (N1) followed by the number of participants in the first round of augmentation (N2), the number of participants in the second round of augmentation (N3), etc.}

\item{pmax}{refers to the maximum p-value in the data collected thusfar such that the researcher would augment the sample with additional participants.}

\item{pdesired}{refers to the desired Type I error rate (typically .05).}

\item{slices}{refers to the number of slices to divide the probability distribution (higher numbers of slices will make the calculations slower but more precise).}

\item{tails}{refers to whether the tests are one- or two-tailed.}
}
\description{
This function calculates pcrit, a statistic discussed in Sagarin, Ambler, and Lee's (2014, Perspectives on Psychological Science, 9, 293-304) "An Ethical Approach to Peeking at Data".
}
\examples{
pcrit(c(200,150,50), pmax = .2, pdesired = .05, slices = 1000, tails = 2)
pcrit()
}
